+++ packages/adhere/changelog/CHANGELOG.md
@@ -11,6 +11,7 @@
 * **adhere-ui-antdformitem**
   - 加入adhere-ui-antdformitem代码
   - 加入Input和TextArea组件
+  - 加入所有Antd组件的默认操作
 * **adhere-ui-searchtable**
   - 加入ProSearchTable
   - 加入editor功能